requirements.txt
# Core dependencies - strictly pinned versions for stability
python-telegram-bot==20.7
google-generativeai==0.3.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.31.0
easyocr==1.7.1
Pillow==10.2.0
google-api-python-client==2.118.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
numpy==1.24.3
flask==2.3.3
flask-cors==4.0.0

# Database and caching
redis==5.0.1
#psycopg2-binary==2.9.9  # For PostgreSQL support (optional)

# Security and validation
cryptography==41.0.7
pydantic==2.5.2
validators==0.22.0
bleach==6.1.0  # For HTML sanitization

# Monitoring and logging
sentry-sdk==1.39.2  # Error tracking
structlog==23.2.0   # Structured logging
prometheus-client==0.19.0  # Metrics

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For async HTTP testing

# Development tools
black==23.11.0  # Code formatting
flake8==6.1.0   # Linting
mypy==1.7.1     # Type checking
isort==5.12.0   # Import sorting

# Production server
gunicorn==21.2.0
uvicorn==0.24.0  # ASGI server

# Performance and optimization
cachetools==5.3.2
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Additional utilities
click==8.1.7  # CLI framework
python-magic==0.4.27  # File type detection
chardet==5.2.0  # Character encoding detection
urllib3==2.1.0  # HTTP library (specific version for security)
certifi==2023.11.17  # SSL certificates

# Optional cloud services
#boto3==1.34.3  # AWS SDK (optional)
#azure-storage-blob==12.19.0  # Azure Blob Storage (optional)

# API rate limiting
slowapi==0.1.9  # FastAPI rate limiting
limits==3.7.0   # Rate limiting library